# Conflicts:
#	src/main/java/it/pagopa/pn/commons/utils/ClientAspectLogging.java
#	src/test/java/it/pagopa/pn/commons/configs/lollipop/PNHttpVerifierConfigurationTest.java
#	src/test/java/it/pagopa/pn/commons/utils/ClientAspectTest.java
#	src/test/java/it/pagopa/pn/national/registries/generated/openapi/msclient/downstream/inad/v1/api/TestApi.java

+++ src/main/java/it/pagopa/pn/commons/log/dto/metrics/Metric.java
@@ -10,7 +10,7 @@ import org.apache.commons.lang3.StringUtils;
 @AllArgsConstructor
 public class Metric {
     private String name;
-    private int value;
+    private long value;
 
     public String toJson() { return StringUtils.EMPTY; }
